backend/requirements-llm.txt
# Install after core deps; may require CMake / Xcode CLI tools to build from source.
llama-cpp-python>=0.3.1
